requirements.txt
annotated-types==0.7.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
google-auth==2.32.0
idna==3.7
inquirerpy==0.3.4
kubernetes==26.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
pfzy==0.3.4
prompt_toolkit==3.0.47
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyhelm3==0.4.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
typer==0.12.3
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
kfp==2.8.0
kfp-kubernetes==1.3.0
numpy==2.1.2
pandas==2.2.3